Code reorg: isolate liquid transformation code
