+++ src/liquidlogicclassic.h
@@ -19,29 +19,44 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 
 #pragma once
 
+#include "liquidlogic.h"
 #include "util/container.h"
 #include "irrlichttypes_bloated.h"
 
-class NodeDefManager;
+class ServerEnvironment;
+class IGameDef;
+class INodeDefManager;
 class Map;
 class MapBlock;
+class MapNode;
 
-class ReflowScan {
+
+class LiquidLogicClassic: public LiquidLogic {
 public:
-	ReflowScan(Map *map, const NodeDefManager *ndef);
-	void scan(MapBlock *block, UniqueQueue<v3s16> *liquid_queue);
-
+	LiquidLogicClassic(Map *map, IGameDef *gamedef);
+	void addTransforming(v3s16 p);
+	void scanBlock(MapBlock *block);
+	void scanVoxelManip(MMVManip *vm, v3s16 nmin, v3s16 nmax);
+	void scanVoxelManip(UniqueQueue<v3s16> *liquid_queue,
+		MMVManip *vm, v3s16 nmin, v3s16 nmax);
+	void transform(std::map<v3s16, MapBlock*> &modified_blocks,
+		ServerEnvironment *env);
+	void addTransformingFromData(BlockMakeData *data);
 private:
 	MapBlock *lookupBlock(int x, int y, int z);
 	bool isLiquidFlowableTo(int x, int y, int z);
 	bool isLiquidHorizontallyFlowable(int x, int y, int z);
+	bool isLiquidHorizontallyFlowable(MMVManip *vm, u32 vi, v3s16 em);
 	void scanColumn(int x, int z);
 
 private:
-	Map *m_map = nullptr;
-	const NodeDefManager *m_ndef = nullptr;
+	UniqueQueue<v3s16> m_liquid_queue;
+
+	u32 m_unprocessed_count = 0;
+	bool m_queue_size_timer_started = false;
+	u64 m_inc_trending_up_start_time = 0; // milliseconds
+
 	v3s16 m_block_pos, m_rel_block_pos;
-	UniqueQueue<v3s16> *m_liquid_queue = nullptr;
 	MapBlock *m_lookup[3 * 3 * 3];
 	u32 m_lookup_state_bitset;
 };
