+++ src/map.h
@@ -46,6 +46,7 @@ class IGameDef;
 class IRollbackManager;
 class EmergeManager;
 class ServerEnvironment;
+class LiquidLogic;
 struct BlockMakeData;
 
 /*
@@ -278,6 +279,9 @@ public:
 	void transforming_liquid_add(v3s16 p);
 
 	bool isBlockOccluded(MapBlock *block, v3s16 cam_pos_nodes);
+
+	inline LiquidLogic * getLiquidLogic() { return m_liquid_logic; }
+
 protected:
 	friend class LuaVoxelManip;
 
@@ -293,9 +297,6 @@ protected:
 	MapSector *m_sector_cache = nullptr;
 	v2s16 m_sector_cache_p;
 
-	// Queued transforming water nodes
-	UniqueQueue<v3s16> m_transforming_liquid;
-
 	// This stores the properties of the nodes on the map.
 	const NodeDefManager *m_nodedef;
 
@@ -305,11 +306,7 @@ protected:
 		float step, float stepfac, float start_offset, float end_offset,
 		u32 needed_count);
 
-private:
-	f32 m_transforming_liquid_loop_count_multiplier = 1.0f;
-	u32 m_unprocessed_count = 0;
-	u64 m_inc_trending_up_start_time = 0; // milliseconds
-	bool m_queue_size_timer_started = false;
+	LiquidLogic *m_liquid_logic;
 };
 
 /*
@@ -472,6 +469,8 @@ public:
 	// This is much faster with big chunks of generated data
 	void blitBackAll(std::map<v3s16, MapBlock*> * modified_blocks,
 		bool overwrite_generated = true);
+
+	inline LiquidLogic * getLiquidLogic() { return m_map->getLiquidLogic(); }
 
 	bool m_is_dirty = false;
